added limits for parsing

+++ app/Product.php
@@ -13,17 +13,29 @@ class Product extends Model
         return $this->hasMany('App\ProductProperty');
     }
 
-    public static function createRecord($attributes)
+    /**
+     * Decide either create, or update or do nothing with parsed data.
+     *
+     * @param $attributes
+     * @param $can_update
+     * @return array|void
+     */
+    public static function manageRecords($attributes, $can_update)
     {
         $product = Product::where('name', $attributes['name'])->first();
 
-        if ($product)
-        {
+        $status = 'created';
+
+        if ($product && $can_update) {
             $product->delete();
+
+            $status = 'updated';
+        } elseif ($product && !$can_update) {
+            return;
         }
 
         $product = Product::create($attributes);
 
-        return $product;
+        return ['product' => $product, 'status' => $status];
     }
 }
